pyproject.toml
[tool.poetry]
name = "project_name"
version = "0.1.1"
description = "Django project template"
authors = ["Louis Agyapong <agyapong.louis@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10.0"
Django = "^4.0"
django-debug-toolbar = "^3.2.4"
isort = "^5.7.0"
djangorestframework = "^3.13.1"
flake8 = "^4.0.1"
dj-database-url = "^0.5.0"
python-decouple = "^3.5"
pytest-django = "^4.5.2"

[tool.poetry.dev-dependencies]
ipython = "^7.30.1"
django-extensions = "^3.1.0"
black = "^21.2b0"
mypy = "^0.930"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
